Issue #659: Update handling for screened comments with parents
The success messages for screened comments with parent comments left by
a logged-in user now mention that the author of the parent comment can
see the comment as well as the author and/or the journal/community owners.

Success messages for screened comments with no parent comments left by
a logged-in user no longer mention a possible parent comment.

This logic is implemented as added conditionals in `talkpost_do.bml`.

The "View your comment" link is now consistent throughout all
comment success messages.

Fixes #659.